requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
cfgv==3.3.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==39.0.2
daphne==4.0.0
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
flake8==6.0.0
hyperlink==21.0.0
identify==2.5.18
idna==3.4
incremental==22.10.0
mccabe==0.7.0
msgpack==1.0.5
nodeenv==1.7.0
platformdirs==3.1.0
pre-commit==3.1.1
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
virtualenv==20.20.0
zope.interface==5.5.2
